fix: SDK stubs for healing/missiles, joinable missile entries, clippy, README
Final-review fix wave for M10: hand-transcribe HealingOut/MissilesOut into
the Node/Python SDK type stubs (present in the schema, missing from both),
add an account field to PlayerMissilesOut so missiles.players[] entries can
be joined back to players[] by account, drop an unnecessary same-type cast
flagged by clippy, and bring the README's usage/view lists and milestones
up to date with M10's healing/missiles/enemy-filter/u32-id work.

+++ crates/axilog-core/src/evtc/ext_healing.rs
@@ -298,7 +298,7 @@ mod tests {
         let e = registration_event(HEALING_SIGNATURE, 2);
         let r = decode_registration(&e).unwrap();
         assert_ne!(r.signature as u64, r.revision as u64);
-        assert_ne!(r.signature, HEALING_SIGNATURE as u32 >> 8, "must not be a shifted mis-decode");
+        assert_ne!(r.signature, HEALING_SIGNATURE >> 8, "must not be a shifted mis-decode");
     }
 
     #[test]
